replayer: remove the debug message in the Replayer module
Logging the debug message in the Replayer module slows down the pipeline
significanty. This commit removes the debug message.

+++ core/drivers/pcap_reader.h
@@ -28,8 +28,8 @@ class PCAPReader final : public Port {
   // The module's pcap file handler
   pcap_t *pcap_handle_ = nullptr;
   // Initial time origin
-  uint32_t init_tsec_;
-  uint32_t init_tusec_;
+  uint64_t init_tsec_;
+  uint64_t init_tusec_;
   uint64_t global_init_ts_;
 };
